backend/requirements.txt
deta
fastapi
pydantic
python-dotenv
uvicorn